# packages/schemas/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HK Open Data Self-hosted Runtime",
    "version": "1.0.0-draft",
    "description": "Normative v0 REST surface. Public status operations are explicitly marked; all other operations require an audience-bound bearer token."
  },
  "servers": [
    {"url": "/", "description": "Current environment; deployment supplies the origin"}
  ],
...
  "paths": {
    "/v1/sources": {
      "get": {
        "operationId": "listSources",
        "summary": "List visible sources",
        "tags": ["Sources"],
        "security": [{}, {"bearerAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/Cursor"},
          {"$ref": "#/components/parameters/Limit"},
          {"name": "project", "in": "query", "schema": {"type": "string", "pattern": "^P[0-9]{2}$"}},
          {"name": "authority_class", "in": "query", "schema": {"type": "string"}},
          {"name": "freshness_status", "in": "query", "schema": {"type": "string"}},
          {"name": "approval_status", "in": "query", "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Visible source summaries", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SourcePage"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"}
        }
      }
    },
    "/v1/sources/{source_id}": {
      "get": {
        "operationId": "getSource",
        "summary": "Get one source and its current health",
        "tags": ["Sources"],
        "security": [{}, {"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/SourceId"}, {"$ref": "#/components/parameters/IfMatch"}],
        "responses": {
          "200": {"description": "Source definition and current health", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SourceDetail"}}}},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
    "/v1/source-records": {
      "get": {
        "operationId": "querySourceRecords",
        "summary": "Query authorized source records",
        "tags": ["Records"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/Cursor"},
          {"$ref": "#/components/parameters/Limit"},
          {"name": "source_id", "in": "query", "schema": {"type": "string"}},
          {"name": "observed_from", "in": "query", "schema": {"type": "string", "format": "date-time"}},
          {"name": "observed_to", "in": "query", "schema": {"type": "string", "format": "date-time"}},
          {"name": "published_from", "in": "query", "schema": {"type": "string", "format": "date-time"}},
          {"name": "published_to", "in": "query", "schema": {"type": "string", "format": "date-time"}},
          {"name": "language", "in": "query", "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Authorized source records", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SourceRecordPage"}}}},
          "401": {"$ref": "#/components/responses/Unauthorized"}
        }
      }
    },
    "/v1/source-records/{source_record_id}": {
      "get": {
        "operationId": "getSourceRecord",
        "summary": "Get one source record with lineage",
        "tags": ["Records"],
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "source_record_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {
          "200": {"description": "One source record with lineage", "content": {"application/json": {"schema": {"$ref": "./source_record.schema.json"}}}},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
...
    "/v1/status/summary": {
      "get": {
        "operationId": "getPublicStatusSummary",
        "summary": "Get the reviewed public status summary",
        "tags": ["Quality"],
        "security": [{}, {"bearerAuth": []}],
        "parameters": [
          {"name": "project", "in": "query", "schema": {"type": "string", "pattern": "^P[0-9]{2}$"}},
          {"name": "provider", "in": "query", "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Reviewed public status summary", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/StatusSummary"}}}}
        }
      }
    },
    "/v1/monitor-targets": {
      "get": {
        "operationId": "listMonitorTargets",
        "summary": "List visible monitor targets",
        "tags": ["Quality"],
        "security": [{}, {"bearerAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/Cursor"},
          {"$ref": "#/components/parameters/Limit"},
          {"name": "provider", "in": "query", "schema": {"type": "string"}},
          {"name": "outcome", "in": "query", "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Public target views; bearer tokens may reveal authorized private targets", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MonitorTargetPage"}}}}
        }
      }
    },
    "/v1/monitor-targets/{monitor_id}": {
      "get": {
        "operationId": "getMonitorTarget",
        "summary": "Get a monitor target and reviewed history",
        "tags": ["Quality"],
        "security": [{}, {"bearerAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/MonitorId"},
          {"name": "history_limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 100, "default": 20}}
        ],
        "responses": {
          "200": {"description": "Target and reviewed observation history", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MonitorTargetDetail"}}}},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
    "/v1/incidents": {
      "get": {
        "operationId": "listIncidents",
        "summary": "List visible incidents",
        "tags": ["Quality"],
        "security": [{}, {"bearerAuth": []}],
        "parameters": [
          {"$ref": "#/components/parameters/Cursor"},
          {"$ref": "#/components/parameters/Limit"},
          {"name": "status", "in": "query", "schema": {"type": "string"}},
          {"name": "severity", "in": "query", "schema": {"type": "string"}},
          {"name": "source_id", "in": "query", "schema": {"type": "string"}},
          {"name": "opened_from", "in": "query", "schema": {"type": "string", "format": "date-time"}},
          {"name": "opened_to", "in": "query", "schema": {"type": "string", "format": "date-time"}}
        ],
        "responses": {
          "200": {"description": "Reviewed public incidents or authorized private incidents", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/IncidentPage"}}}}
        }
      }
    },
    "/v1/incidents/{incident_id}": {
      "get": {
        "operationId": "getIncident",
        "summary": "Get one visible incident",
        "tags": ["Quality"],
        "security": [],
        "parameters": [{"$ref": "#/components/parameters/IncidentId"}],
        "responses": {
          "200": {"description": "Incident view filtered by caller authorization", "content": {"application/json": {"schema": {"$ref": "./incident.schema.json"}}}},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT", "description": "Audience-bound token issued by the deployment's approved OIDC provider."}
    },
    "parameters": {
      "Cursor": {"name": "cursor", "in": "query", "schema": {"type": "string"}},
      "Limit": {"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 200, "default": 50}},
      "SourceId": {"name": "source_id", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^(HKAPI|EXT)-[0-9]{3}$"}},
      "MonitorId": {"name": "monitor_id", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^P14-M[0-9]{3}$"}},
      "IncidentId": {"name": "incident_id", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^INC-[0-9]{4}-[0-9]{6}$"}},
      "IfMatch": {"name": "If-Match", "in": "header", "required": true, "schema": {"type": "string"}, "description": "Current audit version for optimistic concurrency."}
    },
    "responses": {
      "Unauthorized": {"description": "Missing or invalid token", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
...
      "NotFound": {"description": "Object absent or hidden from caller", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}},
...
    },
    "schemas": {
      "PageMeta": {
        "type": "object",
        "required": ["next_cursor"],
        "properties": {"next_cursor": {"type": ["string", "null"]}},
        "additionalProperties": false
      },
      "SourceSummary": {
        "type": "object",
        "required": ["source_id", "operating_profile", "name", "provider", "authority_class", "approval_status", "freshness_status", "last_success_at", "version"],
        "properties": {
          "source_id": {"type": "string"},
          "catalogue_id": {"type": "string", "pattern": "^(official|external|mcp):[a-z0-9][a-z0-9-]*$"},
          "catalogue_verified_at": {"type": "string", "format": "date"},
          "terms_evidence_state": {"enum": ["not-reviewed", "official-terms-linked", "restriction-identified", "ambiguity-identified", "provider-confirmation-recorded"]},
          "operating_profile": {"enum": ["catalogue", "observe", "fabric"]},
          "source_group_id": {"type": ["string", "null"]},
          "name": {"type": "string"},
          "provider": {"type": "string"},
          "authority_class": {"type": "string"},
          "approval_status": {"type": "string"},
          "freshness_status": {"type": "string"},
          "last_success_at": {"type": ["string", "null"], "format": "date-time"},
          "limitations": {"type": "array", "items": {"type": "string"}},
          "documentation_url": {"type": ["string", "null"], "format": "uri"},
          "cadence": {"type": "string"},
          "approved_uses": {"type": "array", "items": {"type": "string"}},
          "version": {"type": "integer", "minimum": 1}
        },
        "additionalProperties": false
      },
      "SourceDetail": {
        "allOf": [
          {"$ref": "#/components/schemas/SourceSummary"},
          {"type": "object", "required": ["documentation_url", "cadence", "approved_uses"]}
        ]
      },
      "SourcePage": {"type": "object", "required": ["items", "page"], "properties": {"items": {"type": "array", "items": {"$ref": "#/components/schemas/SourceSummary"}}, "page": {"$ref": "#/components/schemas/PageMeta"}}, "additionalProperties": false},
      "SourceRecordPage": {"type": "object", "required": ["items", "page"], "properties": {"items": {"type": "array", "items": {"$ref": "./source_record.schema.json"}}, "page": {"$ref": "#/components/schemas/PageMeta"}}, "additionalProperties": false},
...
      "MonitorTargetSummary": {
        "type": "object",
        "required": ["monitor_id", "source_id", "provider", "name", "outcome", "last_checked_at", "public_visibility", "baseline_version"],
        "properties": {
          "monitor_id": {"type": "string"},
          "source_id": {"type": "string"},
          "provider": {"type": "string"},
          "name": {"type": "string"},
          "outcome": {"type": "string"},
          "last_checked_at": {"type": ["string", "null"], "format": "date-time"},
          "public_visibility": {"enum": ["public", "private", "pending_review", "private_until_review"]},
          "baseline_version": {"type": ["integer", "null"], "minimum": 1},
          "activation_status": {"enum": ["specified_pending_approval", "approved", "suspended", "retired"]},
          "operator_identity": {"type": ["string", "null"]},
          "rule_version": {"type": ["string", "null"]},
          "version": {"type": "integer", "minimum": 1}
        },
        "additionalProperties": false
      },
      "MonitorTargetPage": {"type": "object", "required": ["items", "page"], "properties": {"items": {"type": "array", "items": {"$ref": "#/components/schemas/MonitorTargetSummary"}}, "page": {"$ref": "#/components/schemas/PageMeta"}}, "additionalProperties": false},
      "MonitorTargetDetail": {"type": "object", "required": ["target", "observations", "incidents"], "properties": {"target": {"$ref": "#/components/schemas/MonitorTargetSummary"}, "observations": {"type": "array", "items": {"$ref": "./monitor_observation.schema.json"}}, "incidents": {"type": "array", "items": {"$ref": "./incident.schema.json"}}}, "additionalProperties": false},
      "IncidentPage": {"type": "object", "required": ["items", "page"], "properties": {"items": {"type": "array", "items": {"$ref": "./incident.schema.json"}}, "page": {"$ref": "#/components/schemas/PageMeta"}}, "additionalProperties": false},
      "StatusSummary": {
        "type": "object",
        "required": ["operating_profile", "generated_at", "overall", "counts", "current_incidents"],
        "properties": {
          "operating_profile": {"enum": ["catalogue", "observe", "fabric"]},
          "generated_at": {"type": "string", "format": "date-time"},
          "overall": {"enum": ["operational", "degraded", "major_outage", "unknown"]},
          "counts": {"type": "object", "additionalProperties": {"type": "integer", "minimum": 0}},
          "current_incidents": {"type": "array", "items": {"$ref": "./incident.schema.json"}}
        },
        "additionalProperties": false
      },
...
      "Error": {
        "type": "object",
        "required": ["code", "message", "retryable", "correlation_id"],
        "properties": {
          "code": {"type": "string"},
          "message": {"type": "string"},
          "retryable": {"type": "boolean"},
          "correlation_id": {"type": "string"}
        },
        "additionalProperties": false
      }
    }
  }
}